Enable and fix most warnings (Win only)
This also creates a wrapper for gtest.  IntelliSense doesn't look at the
cmake files for warnings, so we can't suppress warnings from gtest.  So
this creates a wrapper to suppress them using #pragma.

+++ src/public/value.cc
@@ -85,14 +85,16 @@ bool Value::operator==(const Value& other) const {
       return std::get<uint64_t>(value_) == std::get<uint64_t>(other.value_);
     } else if (other_type == ValueType::SignedInt) {
       return std::get<int64_t>(other.value_) >= 0 &&
-             std::get<uint64_t>(value_) == std::get<int64_t>(other.value_);
+             std::get<uint64_t>(value_) ==
+                 static_cast<uint64_t>(std::get<int64_t>(other.value_));
     } else if (other_type == ValueType::Double) {
       return std::get<uint64_t>(value_) == std::get<double>(other.value_);
     }
   } else if (this_type == ValueType::SignedInt) {
     if (other_type == ValueType::UnsignedInt) {
       return std::get<int64_t>(value_) >= 0 &&
-             std::get<int64_t>(value_) == std::get<uint64_t>(other.value_);
+             static_cast<uint64_t>(std::get<int64_t>(value_)) ==
+                 std::get<uint64_t>(other.value_);
     } else if (other_type == ValueType::SignedInt) {
       return std::get<int64_t>(value_) == std::get<int64_t>(other.value_);
     } else if (other_type == ValueType::Double) {
